Add argument min_challenge
Add argument min_challenge that like max_challenge sets a lower bound on
the number of challenges.  With min_challenge it becomes possible to deny
authentication when the radius server sends no challenge.

This is used to require TOTP when using Microsoft Azure via Network Policy
Server as the radius server. Azure and NPS does not support demanding one
particular authentication metod. This patch is a workaround for that.